- fix msvc warnings. - introduce ssize_t (for signed size_t type). - new jshost -w option: report warnings as errors. - fix Blob. substr, substring and lastIndexOf functions. - jssqlite: add SqliteError const error names. - jssqlite: Step throw an exception on SQLITE_MISUSE error. - minor other modifications.

+++ src/jscrypt/cipher.cpp
@@ -197,7 +197,7 @@ DEFINE_CONSTRUCTOR() {
 		case mode_ecb: {
 			pv->symmetric_XXX = JS_malloc(cx, sizeof(symmetric_ECB));
 			JL_S_ASSERT_ALLOC( pv->symmetric_XXX );
-			JL_S_ASSERT( keyLength >= (unsigned)cipher->min_key_length && keyLength <= (unsigned)cipher->max_key_length, "Invalid key length (need [%d,%d]  bytes)", cipher->min_key_length, cipher->max_key_length );
+			JL_S_ASSERT( keyLength >= (size_t)cipher->min_key_length && keyLength <= (size_t)cipher->max_key_length, "Invalid key length (need [%d,%d]  bytes)", cipher->min_key_length, cipher->max_key_length );
 //			JL_S_ASSERT( IV == NULL, "Initialization vector is invalid for this mode." );
 			JL_S_ASSERT( optarg == NULL, "invalid 'arg' argument for this mode." );
 			err = ecb_start( cipherIndex, (unsigned char *)key, (int)keyLength, numRounds, (symmetric_ECB *)pv->symmetric_XXX );
@@ -206,7 +206,7 @@ DEFINE_CONSTRUCTOR() {
 		case mode_cfb: {
 			pv->symmetric_XXX = JS_malloc(cx, sizeof(symmetric_CFB));
 			JL_S_ASSERT_ALLOC( pv->symmetric_XXX );
-			JL_S_ASSERT( keyLength >= (unsigned)cipher->min_key_length && keyLength <= (unsigned)cipher->max_key_length, "Invalid key length (need [%d,%d]  bytes)", cipher->min_key_length, cipher->max_key_length );
+			JL_S_ASSERT( keyLength >= (size_t)cipher->min_key_length && keyLength <= (size_t)cipher->max_key_length, "Invalid key length (need [%d,%d]  bytes)", cipher->min_key_length, cipher->max_key_length );
 //			JL_S_ASSERT( IVLength == cipher->block_length, "This cipher require a IV length of %d", cipher->block_length );
 			JL_S_ASSERT( optarg == NULL, "invalid 'arg' argument for this mode." );
 			err = cfb_start( cipherIndex, (unsigned char *)IV, (unsigned char *)key, (int)keyLength, numRounds, (symmetric_CFB *)pv->symmetric_XXX );
@@ -215,7 +215,7 @@ DEFINE_CONSTRUCTOR() {
 		case mode_ofb: {
 			pv->symmetric_XXX = JS_malloc(cx, sizeof(symmetric_OFB));
 			JL_S_ASSERT_ALLOC( pv->symmetric_XXX );
-			JL_S_ASSERT( keyLength >= (unsigned)cipher->min_key_length && keyLength <= (unsigned)cipher->max_key_length, "Invalid key length (need [%d,%d]  bytes)", cipher->min_key_length, cipher->max_key_length );
+			JL_S_ASSERT( keyLength >= (size_t)cipher->min_key_length && keyLength <= (size_t)cipher->max_key_length, "Invalid key length (need [%d,%d]  bytes)", cipher->min_key_length, cipher->max_key_length );
 //			JL_S_ASSERT( IVLength == cipher->block_length, "This cipher require a IV length of %d", cipher->block_length );
 			JL_S_ASSERT( optarg == NULL, "invalid 'arg' argument for this mode." );
 			err = ofb_start( cipherIndex, (unsigned char *)IV, (unsigned char *)key, (int)keyLength, numRounds, (symmetric_OFB *)pv->symmetric_XXX );
@@ -224,7 +224,7 @@ DEFINE_CONSTRUCTOR() {
 		case mode_cbc: {
 			pv->symmetric_XXX = JS_malloc(cx, sizeof(symmetric_CBC));
 			JL_S_ASSERT_ALLOC( pv->symmetric_XXX );
-			JL_S_ASSERT( keyLength >= (unsigned)cipher->min_key_length && keyLength <= (unsigned)cipher->max_key_length, "Invalid key length (need [%d,%d]  bytes)", cipher->min_key_length, cipher->max_key_length );
+			JL_S_ASSERT( keyLength >= (size_t)cipher->min_key_length && keyLength <= (size_t)cipher->max_key_length, "Invalid key length (need [%d,%d]  bytes)", cipher->min_key_length, cipher->max_key_length );
 //			JL_S_ASSERT( IVLength == cipher->block_length, "This cipher require a IV length of %d", cipher->block_length );
 			JL_S_ASSERT( optarg == NULL, "invalid 'arg' argument for this mode." );
 			err = cbc_start( cipherIndex, (unsigned char *)IV, (unsigned char *)key, (int)keyLength, numRounds, (symmetric_CBC *)pv->symmetric_XXX );
@@ -233,7 +233,7 @@ DEFINE_CONSTRUCTOR() {
 		case mode_ctr: {
 			pv->symmetric_XXX = JS_malloc(cx, sizeof(symmetric_CTR));
 			JL_S_ASSERT_ALLOC( pv->symmetric_XXX );
-			JL_S_ASSERT( keyLength >= (unsigned)cipher->min_key_length && keyLength <= (unsigned)cipher->max_key_length, "Invalid key length (need [%d,%d]  bytes)", cipher->min_key_length, cipher->max_key_length );
+			JL_S_ASSERT( keyLength >= (size_t)cipher->min_key_length && keyLength <= (size_t)cipher->max_key_length, "Invalid key length (need [%d,%d]  bytes)", cipher->min_key_length, cipher->max_key_length );
 //			JL_S_ASSERT( IVLength == cipher->block_length, "This cipher require a IV length of %d", cipher->block_length );
 			JL_S_ASSERT( optarg == NULL, "invalid 'arg' argument for this mode." );
 			err = ctr_start( cipherIndex, (unsigned char *)IV, (unsigned char *)key, (int)keyLength, numRounds, CTR_COUNTER_LITTLE_ENDIAN, (symmetric_CTR *)pv->symmetric_XXX );
@@ -242,7 +242,7 @@ DEFINE_CONSTRUCTOR() {
 		case mode_lrw: {
 			pv->symmetric_XXX = JS_malloc(cx, sizeof(symmetric_LRW));
 			JL_S_ASSERT_ALLOC( pv->symmetric_XXX );
-			JL_S_ASSERT( keyLength >= (unsigned)cipher->min_key_length && keyLength <= (unsigned)cipher->max_key_length, "Invalid key length (need [%d,%d]  bytes)", cipher->min_key_length, cipher->max_key_length );
+			JL_S_ASSERT( keyLength >= (size_t)cipher->min_key_length && keyLength <= (size_t)cipher->max_key_length, "Invalid key length (need [%d,%d]  bytes)", cipher->min_key_length, cipher->max_key_length );
 //			JL_S_ASSERT( IVLength == cipher->block_length, "This cipher require a IV length of %d", cipher->block_length );
 			JL_S_ASSERT( optargLength == keyLength, "The tweak length must be %d bytes length (key size)", (int)keyLength );
 			err = lrw_start( cipherIndex, (unsigned char *)IV, (unsigned char *)key, (int)keyLength, (unsigned char *)optarg, numRounds, (symmetric_LRW *)pv->symmetric_XXX );
@@ -251,7 +251,7 @@ DEFINE_CONSTRUCTOR() {
 		case mode_f8: {
 			pv->symmetric_XXX = JS_malloc(cx, sizeof(symmetric_F8));
 			JL_S_ASSERT_ALLOC( pv->symmetric_XXX );
-			JL_S_ASSERT( keyLength >= (unsigned)cipher->min_key_length && keyLength <= (unsigned)cipher->max_key_length, "Invalid key length (need [%d,%d]  bytes)", cipher->min_key_length, cipher->max_key_length );
+			JL_S_ASSERT( keyLength >= (size_t)cipher->min_key_length && keyLength <= (size_t)cipher->max_key_length, "Invalid key length (need [%d,%d]  bytes)", cipher->min_key_length, cipher->max_key_length );
 //			JL_S_ASSERT( IVLength == cipher->block_length, "This cipher require a IV length of %d", cipher->block_length );
 			JL_S_ASSERT( optargLength > 0, "This mode need the salt argument" );
 			err = f8_start( cipherIndex, (unsigned char *)IV, (unsigned char *)key, (int)keyLength, (unsigned char *)optarg, (int)optargLength, numRounds, (symmetric_F8 *)pv->symmetric_XXX );
